+++ idea/testData/refactoring/rename/varInDoWhile/after/test.kt
@@ -1,0 +1,6 @@
+fun test() {
+    do {
+        val message2 = "test"
+        println(message2)
+    } while (message2.isEmpty())
+}